+++ teneto/__init__.py
@@ -4,7 +4,7 @@
 """Teneto is a module with tools for analyzing network patterns in time."""
 #
 __author__ = "William Hedley Thompson (wiheto)"
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 #
 import teneto.utils as utils
 import teneto.networkmeasures as networkmeasures
@@ -12,8 +12,9 @@ import teneto.plot as plot
 import teneto.generatenetwork as generatenetwork
 import teneto.derive as derive
 import teneto.misc as misc
+import teneto.trajectory as trajectory
 #import teneto.communitydetection as communitydetection
-#from teneto.classes import NewTemporalNetwork
+from teneto.classes import NewTemporalNetwork, TenetoBIDS
 #del misc.teneto
 #del communitydetection.static.modularity_based_clustering