+++ teneto/classes/__init__.py
@@ -1,0 +1,2 @@
+from teneto.classes.network import NewTemporalNetwork
+from teneto.classes.bids import TenetoBIDS