+++ teneto/utils/__init__.py
@@ -2,13 +2,13 @@
 Import utility funcitons
 """
 
-from teneto.utils.utils import checkInput
-from teneto.utils.utils import contact2graphlet
+from teneto.utils.utils import checkInput, process_input, gen_nettype
+from teneto.utils.utils import contact2graphlet, graphlet2contact
 from teneto.utils.utils import getDistanceFunction
 from teneto.utils.utils import set_diagonal
-from teneto.utils.utils import gen_nettype
-from teneto.utils.utils import graphlet2contact
-from teneto.utils.utils import process_input
 from teneto.utils.utils import clean_community_indexes
 from teneto.utils.utils import multiple_contacts_get_values
 from teneto.utils.utils import check_distance_funciton_input
+from teneto.utils.utils import binarize, binarize_percent, binarize_magnitude, binarize_rdp
+from teneto.utils.utils import create_traj_ranges
+from teneto.utils.utils import make_parcellation, load_parcellation_coords